SQLliteSession/pyproject.toml
[project]
name = "sqllitesession"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-agents",
]

[tool.uv.sources]
openai-agents = { git = "https://github.com/openai/openai-agents-python", rev = "main" }
